feat(docker): Update docker-compose configuration to specify Dockerfiles for backend and tele_bot services; enhance healthcheck command format for PostgreSQL service. Refactor package.json to streamline start and development scripts, removing legacy server.js file.
