--- src/agendas/delete_agenda.js
@@ -1,10 +1,0 @@
-class DeleteAgenda {
-
-    constructor(bot) {
-        this.bot = bot;
-    }
-
-    deleteAgenda(msg) {}
-}
-
-export default DeleteAgenda;
