--- src/agendas/paid.js
@@ -1,91 +1,0 @@
-// const { updateAgendaPayment } = require("../DB/querys/querys");
-// const allAgendaAsKeyboard = require("../utilities/all_agenda_as_keyboard");
-// const agendaFormatter = require("../utilities/agenda_formatter");
-
-import { updateAgendaPayment } from "../DB/querys/querys.js";
-import allAgendaAsKeyboard from "../utilities/all_agenda_as_keyboard.js";
-import agendaFormatter from "../utilities/agenda_formatter.js";
-
-class paid{
-    constructor(bot){
-        this.bot = bot
-
-        this._selectedAgendaId = '';
-    }
-
-    get selectedAgendaId(){
-        return this._selectedAgendaId;
-    }
-
-    set selectedAgendaId(v){
-        this._selectedAgendaId = v;
-    }
-
-    async paid(msg){
-        await this._getAllAgendas(msg);
-    }
-
-    /**
-     * Retrieves all agendas from the database and generates a keyboard for the user to select who paid.
-     * @param {Message} msg - The Telegram message object.
-     * @return {Promise<void>}
-     */
-    async _getAllAgendas(msg) {
-        // Get all agendas from the database and generate a keyboard
-        const buttons = await allAgendaAsKeyboard(msg.from.id, 'paid');
-        // Create a response message based on whether the keyboard was generated or not
-        const responseMessage = buttons
-            ? 'Por favor, selecione o fatura:\n\nPor favor, selecione o fatura que voce marca:'
-            : 'Voce nao tem nenhuma fatura encontrada';
-
-        // Send the message with the keyboard to the user
-        this.bot.sendMessage(msg.chat.id, responseMessage, {
-            message_thread_id: msg.message_thread_id,
-            reply_markup: buttons ? { inline_keyboard: buttons } : undefined
-        });
-    }
-    async addToDatabase(callbackQuery, data){
-        const msg = callbackQuery.message;
-        const topicId = process.env.PAID_THREAD_ID;
-
-        const updateAgenda = await updateAgendaPayment(this.selectedAgendaId, data, callbackQuery.from.id);
-        if(updateAgenda) {
-            const alert =  `A conta deª *${updateAgenda.title}* no valor de *${updateAgenda.amount}* com vencimento para *${updateAgenda.date}* foi marcada como paga. ✅`;
-            this.bot.editMessageText(alert, {
-                chat_id: msg.chat.id,
-                message_id: msg.message_id,
-                parse_mode: 'Markdown',
-                reply_markup: {
-                    inline_keyboard: [[]],
-                    remove_keyboard: true
-                },
-            });
-
-
-            this.bot.sendMessage(updateAgenda.chatId, alert, {
-                message_thread_id: topicId,
-                message_id: updateAgenda.messageThreadId,
-                parse_mode: 'Markdown',
-
-            })
-        }
-        // await updateAgendaPayment(this.selectedAgendaId, data, (updateAgenda) => {
-        // })
-    }
-
-    async handlePaid(callbackQuery){
-        const userId = callbackQuery.from.id;
-        const data = callbackQuery.data;
-        const chatId = callbackQuery.message.chat.id;
-        const message_id = callbackQuery.message.message_id;
-        if(data.startsWith('paid_')){
-            this.selectedAgendaId = data.split('_')[1];
-
-            await this.addToDatabase(callbackQuery, {
-                isPaid: true
-            });
-        }
-    }
-}
-
-export default paid;
