--- src/utilities/agenda_formatter.js
@@ -1,27 +1,0 @@
-/**
- * Formats the agenda object into a string that can be sent to the user.
- * @param {Object} agenda - The agenda object containing the payment information.
- * @returns {string} The formatted string.
- */
-function agendaFormatter(agenda) {
-    const members = Array.isArray(agenda.members) ? agenda.members : [];
-    const memberLines = members.length > 0
-        ? members
-            .filter((member) => member.isResponsible !== false)
-            .map((member) => `${member.displayName || member.telegramId} ${member.isPaid ? ' ✅' : '  ❌'}`)
-            .join(' \n')
-        : `Pago ${agenda.isPaid ? ' ✅' : '  ❌'}`;
-
-    return `⚠️⚠️ *ATTENTION ${agenda.title.toUpperCase()} BILL* ⚠️⚠️\n\n`  +
-            `R$ ${agenda.amount} \n\n` +
-            `*Vencimento:* ${agenda.date}\n\n` +
-            `Descrição da fatura: ${agenda.description}\n\n` +
-            `⚠️ Favor informar, na descrição do pagamento, o nome da fatura correspondente. Obrigado! ⚠️\n\n` +
-            `Pix Chave (${agenda.bank}): \`${agenda.pix}\`\n\n` +
-            `--------------------------------\n` +
-            `${memberLines}\n` +
-            `--------------------------------\n\n`;   
-}
-
-
-export default agendaFormatter;
